Séparation des forms et de views.py
